moved users utils
