+++ src/pages/api/users/profile/friends/all.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { getAuthFromApi, verifyUser } from "../../../../../utils/auth";
 import { getFriendsInvited, getFriendsInviting } from "../../../../../database/utils/friends";
-import { getUser, filterUsers } from "../../../../../utils/users";
+import { getUser, filterUsers } from "../../../../../database/utils/users";
 
 export default async function Friends(req:NextApiRequest, res:NextApiResponse) {
