+++ src/pages/api/users/profile/friends/index.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { getAuthFromApi, verifyUser } from "../../../../../utils/auth";
-import { filterUsers, getUser } from "../../../../../utils/users";
+import { filterUsers, getUser } from "../../../../../database/utils/users";
 
 export default async function Friends(req:NextApiRequest, res:NextApiResponse) {
